backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.12
python-multipart==0.0.6
cairosvg==2.7.1
aiofiles==23.2.1
python-dotenv==1.0.0
python-pptx